// frontend/src/generated-types/MissionFrame.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Coordinate frame of a mission item.
 */
export type MissionFrame = "global" | "local_ned" | "global_relative_alt" | "local_enu" | "global_int" | "global_relative_alt_int" | "local_offset_ned" | "body_ned" | "global_terrain_alt" | "global_terrain_alt_int";